Fix template path
The previous way worked as long as you use a case-sensitive file system, but since I run docfx on Linux this is not the case for me. Without this change, the docfx theme would not be applied at all (the site would be generated using the default docfx theme only).